requirements.txt
Flask==2.3.0
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.1.0
Flask-Login==0.6.2
Werkzeug==2.3.0
Chart.js==3.9.1
